# openapi.yaml
openapi: 3.0.3
info:
  title: ias_stt — Trap Management API
  version: 1.0.0
  description: |
    Trap management service with deployment tracking, location
    history, and photo upload. Telemetry is ingested from the
    ChirpStack HTTP integration via `POST /api/telemetry/ingest`
    (an optional MQTT listener remains available). Application users obtain
    JWT access and refresh tokens from `/auth/login` and send access tokens in
    the `Authorization` header. A legacy API key remains available for service
    integrations.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/uplinks:
    get:
      tags: [Uplinks]
      summary: List stored tracker uplinks
      description: |
        Returns persisted uplinks from registered tracker devices, newest
        first. The raw payload is excluded from list results; use the detail
        endpoint to retrieve it.
      operationId: listUplinks
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - name: device_eui
          in: query
          schema: { type: string, maxLength: 100 }
          description: Filter by tracker device EUI
        - name: source
          in: query
          schema: { type: string, enum: [http, mqtt] }
          description: Filter by ingestion source
      responses:
        200:
          description: Stored uplinks, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrackerUplink'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
  /api/uplinks/{uplink_id}:
    parameters:
      - name: uplink_id
        in: path
        required: true
        schema: { type: integer }
        description: Uplink primary key ID
    get:
      tags: [Uplinks]
      summary: Get one stored tracker uplink
      description: Includes the original decoded JSON payload.
      operationId: getUplink
      responses:
        200:
          description: Stored uplink details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackerUplinkDetail'
        401:
          $ref: '#/components/responses/401'
        404:
          $ref: '#/components/responses/404'

  # -- Trackers ---------------------------------------------------------------
...
  /api/traps:
    get:
      tags: [Traps]
      summary: List traps
      operationId: listTraps
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - name: status
          in: query
          schema: { type: string, enum: [active, inactive] }
          description: Filter by trap status
      responses:
        200:
          description: A JSON array of traps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trap'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'

    post:
      tags: [Traps]
      summary: Create a trap
      operationId: createTrap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrapCreate'
            example:
              status: active
              trap_id: TRAP-001
              tracker_id: TRK-001
              location: north
              temperature: 23.5
              updated_by: alice
      responses:
        201:
          description: Trap created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trap'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        409:
          $ref: '#/components/responses/409'
  /api/traps/{trap_id}:
    parameters:
      - name: trap_id
        in: path
        required: true
        schema: { type: integer }
        description: Trap primary key ID

    get:
      tags: [Traps]
      summary: Get a trap by ID
      operationId: getTrap
      responses:
        200:
          description: Trap found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trap'
        401:
          $ref: '#/components/responses/401'
        404:
          $ref: '#/components/responses/404'

    put:
      tags: [Traps]
      summary: Update a trap
      description: |
        Requires `updated_by`. Status changes trigger automatic
        deployment create/close; location changes record location
        history under the active deployment.
      operationId: updateTrap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrapUpdate'
            example:
              status: inactive
              location: east
              updated_by: alice
      responses:
        200:
          description: Trap updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trap'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        404:
          $ref: '#/components/responses/404'
        409:
          $ref: '#/components/responses/409'

    delete:
      tags: [Traps]
      summary: Delete a trap
      operationId: deleteTrap
      responses:
        200:
          description: Trap deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
              example:
                message: Trap 1 deleted
        401:
          $ref: '#/components/responses/401'
        404:
          $ref: '#/components/responses/404'

  # -- Deployments -----------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Application users pass the JWT access token as
        `Authorization: Bearer <access-token>`. Legacy service integrations
        may pass the configured API key in the same header.

  # -- Reusable parameters ---------------------------------------------------
  parameters:
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 100, minimum: 0 }
      description: Maximum items per page
    offset:
      name: offset
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
      description: Number of items to skip

  # -- Schemas ---------------------------------------------------------------
  schemas:
    Error:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              properties:
                code: { type: string }
                message: { type: string }
      example:
        error:
          code: AUTHENTICATION_REQUIRED
          message: Authentication is required.
...
    TrackerUplink:
      type: object
      properties:
        id:           { type: integer, readOnly: true }
        device_eui:   { type: string, maxLength: 100 }
        display_name: { type: string, nullable: true }
        received_at:  { type: string, format: date-time }
        source:       { type: string, enum: [http, mqtt] }
        topic:        { type: string, nullable: true }
        latitude:     { type: number, format: float, nullable: true }
        longitude:    { type: number, format: float, nullable: true }
        tilt_status:  { type: string, nullable: true }
        battery:      { type: integer, nullable: true }
      required: [id, device_eui, received_at, source]
    TrackerUplinkDetail:
      allOf:
        - $ref: '#/components/schemas/TrackerUplink'
        - type: object
          properties:
            raw_payload: { type: string }
          required: [raw_payload]
    Trap:
      type: object
      properties:
        id:                { type: integer, readOnly: true }
        status:            { type: string, enum: [active, inactive] }
        trap_id:           { type: string, maxLength: 50 }
        tracker_id:        { type: string, maxLength: 50 }
        location:          { type: string, maxLength: 50, nullable: true }
        door_status:       { type: string, maxLength: 20, nullable: true }
        temperature:       { type: number, format: float, nullable: true }
        notes:             { type: string, maxLength: 255, nullable: true }
        updated_by:        { type: string, maxLength: 50, description: Ignored; derived from the authenticated principal. }
        created_at:        { type: string, format: date-time, readOnly: true }
        updated_at:        { type: string, format: date-time, readOnly: true }
      required:
        - status
        - trap_id
    TrapCreate:
      type: object
      properties:
        status:            { type: string, enum: [active, inactive] }
        trap_id:           { type: string, maxLength: 50 }
        tracker_id:        { type: string, maxLength: 50, description: Device EUI of an existing tracker; may be omitted/empty. }
        location:          { type: string, maxLength: 50 }
        door_status:       { type: string, maxLength: 20 }
        temperature:       { type: number, format: float }
        notes:             { type: string, maxLength: 255 }
        updated_by:        { type: string, maxLength: 50, description: Ignored; derived from the authenticated principal. }
      required: [status, trap_id]
    TrapUpdate:
      type: object
      properties:
        status:            { type: string, enum: [active, inactive] }
        trap_id:           { type: string, maxLength: 50 }
        tracker_id:        { type: string, maxLength: 50 }
        location:          { type: string, maxLength: 50 }
        door_status:       { type: string, maxLength: 20 }
        temperature:       { type: number, format: float }
        notes:             { type: string, maxLength: 255 }
        updated_by:        { type: string, maxLength: 50, description: Ignored; derived from the authenticated principal. }
...
  responses:
    400:
      description: Bad request — missing or invalid fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missingField:
              summary: Missing required field
              value: { error: "Missing required field(s): status" }
            invalidType:
              summary: Invalid limit/offset
              value: { error: "'limit' and 'offset' must be integers" }
    401:
      description: Missing or invalid JWT or service credential
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: AUTHENTICATION_REQUIRED
              message: Authentication is required.
...
    404:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Trap not found
    409:
      description: Conflict — duplicate trap_id
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: trap_id 'TRAP-001' already exists
...